+++ src/main/java/data/ScratchCard.java
@@ -5,7 +5,7 @@ import data.interfaces.Expirable;
 import java.util.Date;
 
 public class ScratchCard extends Product implements Expirable{
-    public ScratchCard(String name, int price, int quantity, Date expiryDate) {
+    public ScratchCard(String name, double price, int quantity, Date expiryDate) {
         super(name, price, quantity, expiryDate);
     }
